Using `eventStoreKey` instead of string for testcases
This is essentially what the tests are dealing with anyway. Also, this
makes it possible to test parsing of serialized `eventStoreKey`s, too
(which is added in this commit).

This commit also slightly changed the signature of `byteCounter´
increment function. This was so that it could be cascaded in testcases.
I'm still trying to wrap my head around the bound functions being bound
to pointers vs. non-pointers. It doesn't feel entirely logical.